% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AI_INDICATOR}
\alias{AI_INDICATOR}
\title{Aleutian Islands (1986-2018)}
\format{
A list containing two data frames (indicator for the full region, indicator by stratum) and a character vector.
\describe{
     \itemize{
       \item FULL_REGION (data frame): Residuals for the full region
         \itemize{
         \item{YEAR}{: Year}
         \item{COMMON_NAME}{: Year}
         \item{mean_wt_resid}{: Mean residual for the full region}
         \item{se_wt_resid}{: Standard error of the indicator for the full region}
         }
       \item STRATUM (data frame): Residuals by stratum
        \itemize{
         \item{YEAR}{: Year}
         \item{COMMON_NAME}{: Species Common name}
         \item{SPECIES_CODE}{: RACE/GAP species code}
         \item{INPFC_STRATUM}{: INPFC Stratum}
         \item{stratum_resid_mean}{: Unweighted stratum mean residual}
         \item{n}{: Sample size for the stratum}
         \item{stratum_resid_sd}{: Standard error of the stratum mean residual}
         \item{weighted_resid_mean}{: Stratum mean Residual weighted in proportion to stratum biomass}
         \item{weighted_resid_mean}{: Standard error of the stratum mean residual weighted in proportion to stratum biomass}
         }
       \item LAST_UPDATE: Date when indicator and data were last updated.
     }
}
}
\source{
\url{https://www.fisheries.noaa.gov/contact/groundfish-assessment-program}
}
\usage{
AI_INDICATOR
}
\description{
Morphometric condition indicators based on residuals from a length-weight regression for Pacific cod, walleye pollock (> 250 mm), walleye pollock (100-250 mm), southern rock sole, arrowtooth flounder, Atka mackerel, northern rockfish, and Pacific ocean perch in the Aleutian Islands.
}
\keyword{datasets}
