% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vast_functions.R
\name{bundle_vast_condition}
\alias{bundle_vast_condition}
\title{Read-in VAST index outputs from .csv}
\usage{
bundle_vast_condition(region, years)
}
\arguments{
\item{region}{Region. AI, BS, or GOA. Character (1L).}

\item{years}{Vector of years to retain in index (i.e. survey years)}
}
\value{
A data frame with condition indicators for all species.
}
\description{
Combines VAST index.csv files from the /results/region/ directory.
}
\examples{
# Combine index.csv files for GOA
bundle_vast_condition(region = "GOA",
years = c(seq(1984, 1999, 3), 
seq(2001, 2022, 2)))
}
