% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lw_residuals.R
\name{calc_lw_residuals}
\alias{calc_lw_residuals}
\title{A function to calculate length-weight residuals}
\usage{
calc_lw_residuals(
  len,
  wt,
  stratum = NA,
  bias.correction = TRUE,
  outlier.rm = FALSE,
  make_diagnostics = FALSE,
  species_code = NA,
  year = NA,
  region = NA
)
}
\arguments{
\item{stratum}{Stratum code for length-weight regression by stratum.}

\item{outlier.rm}{Should outliers be removed using Bonferoni test (cutoff = 0.7)}

\item{make_diagnostics}{Output diagnostic plots and summaries? Default FALSE produces no diagnostics and summaries. If true, species_code and region should be specified.}

\item{species_code}{Species code (provide if make_diagnostics = TRUE)}

\item{year}{Year (provide if make_diagnostics = TRUE)}

\item{region}{Region (provide if make_diagnostics = TRUE)}

\item{length}{Set of individual fish lengths.}

\item{weight}{Corresponding set of individual fish weights.}

\item{bias.correct}{Bias corrected residuals following Brodziak (2012)}
}
\description{
This function makes a log-log regression of length and weight for individual fish and then calculates a residual. A Bonferroni-corrected outlier correction can be applied to remove outliers. Option to use separate covariates for different strata.
}
\references{
Brodziak, J.2012. Fitting length-weight relationships with linear regression using the log-transformed allometric model with bias-correction. Pacific Islands Fish. Sci. Cent., Natl. Mar. Fish. Serv., NOAA, Honolulu, HI 96822-2396. Pacific Islands Fish. Sci. Cent. Admin. Rep. H-12-03, 4 p.
}
\keyword{condition}
\keyword{groundfish}
\keyword{length,}
\keyword{weight,}
