% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_lw_residuals_multimodel.R
\name{calc_lw_residuals_multimodel}
\alias{calc_lw_residuals_multimodel}
\title{Length-weight residuals from best linear regression among candidate models with}
\usage{
calc_lw_residuals_multimodel(
  len,
  wt,
  stratum = NULL,
  sex = NULL,
  day_of_year = NULL,
  year = NULL,
  region = "all",
  species_code = NA,
  bias_correction = TRUE,
  make_diagnostics = TRUE,
  outlier_rm = TRUE,
  include_ci = TRUE,
  multiple_models = TRUE,
  ...
)
}
\arguments{
\item{stratum}{Stratum code for length-weight regression by stratum.}

\item{sex}{Sex code for fish; convertible to a factor.}

\item{day_of_year}{Day of year as an integer.}

\item{year}{Numeric vector for year (provide if make_diagnostics = TRUE)}

\item{region}{Character vector for region (provide if make_diagnostics = TRUE)}

\item{species_code}{Species code (provide if make_diagnostics = TRUE)}

\item{make_diagnostics}{Output diagnostic plots and summaries? Default FALSE produces no diagnostics and summaries. If true, species_code and region should be specified.}

\item{outlier_rm}{Should outliers be removed using Bonferroni test (cutoff = 0.7)}

\item{include_ci}{Include upper and lower estimates for the 95 pct confidence interval of residuals in the output.}

\item{length}{Set of individual fish lengths.}

\item{weight}{Corresponding set of individual fish weights.}

\item{bias_correct}{Bias corrected residuals following Brodziak (2012)}
}
\description{
Fits 1 (no stratum or sex covariates), 4 (either no stratum or no sex covariates), or 15 (stratum and sex covariates) linear regression models to log-weight and log-length with varying combinations of stratum-specific and sex-specific coefficients for allometric slope and intercept terms. The most parsimonious model is selected based on AIC then length-weight residuals are calculated using the best-fit model. Can optionally remove outliers based on a Bonferroni test (outlier_rm = TRUE), bias correct log-weight predictions (bias_correction = TRUE), write regression diagnostics to /output/ directory (make_diagnostics = TRUE), and include confidence intervals for
}
\details{
Brodziak, J.2012. Fitting length-weight relationships with linear regression using the log-transformed allometric model with bias-correction. Pacific Islands Fish. Sci. Cent., Natl. Mar. Fish. Serv., NOAA, Honolulu, HI 96822-2396. Pacific Islands Fish. Sci. Cent. Admin. Rep. H-12-03, 4 p.
}
