% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EBS_INDICATOR}
\alias{EBS_INDICATOR}
\title{Eastern Bering Sea continental shelf (1999-2022)}
\format{
A list containing two data frames (indicator for the full region, indicator by stratum) and a character vector.
\describe{
     \itemize{
       \item FULL_REGION (data frame): Residuals for the full region
         \itemize{
         \item{year}{: Year}
         \item{common_name}{: Year}
         \item{mean_wt_resid}{: Mean residual for the full region}
         \item{se_wt_resid}{: Standard error of the indicator for the full region}
         \item{species_code}{: RACE/GAP species code}
         }
       \item STRATUM (data frame): Residuals by stratum
        \itemize{
         \item{common_name}{: Species Common name}
         \item{year}{: Year}
         \item{stratum}{: Survey Stratum}
         \item{stratum_resid_mean}{: Unweighted stratum mean residual}
         \item{stratum_resid_sd}{: Standard error of the stratum mean residual}
         \item{n}{: Sample size for the stratum}
         \item{weighted_resid_mean}{: Stratum mean Residual weighted in proportion to stratum biomass}
         \item{weighted_resid_se}{: Standard error of the stratum mean residual weighted in proportion to stratum biomass}
         }
       \item LAST_UPDATE: Date when indicator and data were last updated.
     }
}
}
\source{
\url{https://www.fisheries.noaa.gov/contact/groundfish-assessment-program}
}
\usage{
EBS_INDICATOR
}
\description{
Morphometric condition indicators based on length-weight regressions for Pacific cod, walleye pollock (> 250 mm), walleye pollock (100-250 mm), northern rock sole, arrowtooth flounder, flathead sole, Alaska plaice, and yellowfin sole in the eastern Bering Sea.
}
\keyword{datasets}
