% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_condition_data}
\alias{get_condition_data}
\title{Get data for stratum-weighted and VAST methods}
\usage{
get_condition_data(channel = NULL)
}
\arguments{
\item{channel}{RODBC channel}
}
\description{
Retrieve length-weight, CPUE, and stratum-weighted biomass data from RACEBASE for the AI, GOA, EBS, and NBS using queries included in the akfishcondition package. Write output as csvs to /data/ directory.
}
