% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_connected}
\alias{get_connected}
\title{Connect to Oracle using RODBC}
\usage{
get_connected(schema = NA)
}
\arguments{
\item{schema}{Oracle dsn as a character vector.}
}
\description{
This function connects to oracle using RODBC.
}
