% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCOD_ESP}
\alias{PCOD_ESP}
\title{Pacific cod Ecosystem and Socioeconomic Profile (ESP) indicator}
\format{
Eight data frames (indicator for the full region and indicator by stratum for the EBS, NBS, GOA, AI) and a character vector.
\describe{
     \itemize{
       \item FULL_REGION_* (data frame): Residuals for the full region
         \itemize{
         \item{YEAR}{: Year}
         \item{COMMON_NAME}{: Year}
         \item{mean_wt_resid}{: Mean residual for the full region}
         \item{se_wt_resid}{: Standard error of the indicator for the full region}
         }
       \item STRATUM_* (data frame): Residuals by stratum
        \itemize{
         \item{YEAR}{: Year}
         \item{COMMON_NAME}{: Species Common name}
         \item{SPECIES_CODE}{: RACE/GAP species code}
         \item{INPFC_STRATUM}{: INPFC Stratum}
         \item{stratum_resid_mean}{: Unweighted stratum mean residual}
         \item{n}{: Sample size for the stratum}
         \item{stratum_resid_sd}{: Standard error of the stratum mean residual}
         \item{weighted_resid_mean}{: Stratum mean Residual weighted in proportion to stratum biomass}
         \item{weighted_resid_mean}{: Standard error of the stratum mean residual weighted in proportion to stratum biomass}
         }
       \item LAST_UPDATE: Date when indicator and data were last updated.
     }
}
}
\source{
\url{https://www.fisheries.noaa.gov/contact/groundfish-assessment-program}
}
\usage{
PCOD_ESP
}
\description{
Morphometric condition indicators based on residuals from a length-weight regression for adult and juvenile Pacific cod. Separate indicator for each region.
}
\keyword{datasets}
