% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anomaly_timeseries.R
\name{plot_anomaly_timeseries}
\alias{plot_anomaly_timeseries}
\title{Make condition time series plot}
\usage{
plot_anomaly_timeseries(
  x,
  region,
  fill_color = "#0085CA",
  var_y_name = "mean_wt_resid",
  var_y_se_name = "se_wt_resid",
  var_x_name = "year",
  y_title = "Length-weight residual (ln(g))",
  format_for = "rmd",
  plot_type = "point",
  set_intercept = NULL
)
}
\arguments{
\item{x}{Input data.frame}

\item{region}{Region. AI, BS, or GOA. Character (1L).}

\item{fill_color}{Fill color to use for points. Character (1L).}

\item{var_y_name}{Name of the y variable in the data.frame. Character (1L).}

\item{var_y_se_name}{Name of the standard error for the y variable. Character (1L).}

\item{var_x_name}{Name of the x (time variable). Character (1L).}

\item{y_title}{Y-axis title. Character (1L).}

\item{format_for}{"rmd" or "png"}

\item{set_intercept}{Intercept to use for plots. Numeric (1L). 0 for residual, 1 for VAST relative condition, NULL for VAST a}
}
\value{
A ggplot object of the time series
}
\description{
Make condition time series plot
}
\examples{
\dontrun{
# EBS anomaly timeseries plot

names(EBS_INDICATOR$FULL_REGION)

plot_anomaly_timeseries(x = EBS_INDICATOR$FULL_REGION,
fill_color = "#0085CA", 
var_y_name = "mean_wt_resid", 
var_y_se_name = "se_wt_resid",
var_x_name = "year",
y_title = "Length-weight residual (ln(g))",
format_for = "png")
}
}
