% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_bar.R
\name{plot_species_bar}
\alias{plot_species_bar}
\title{Make single species bar plots}
\usage{
plot_species_bar(
  x,
  region,
  var_x_name = "year",
  var_y_name,
  var_y_se_name,
  fill_color = "#0085CA",
  y_title = "Length-weight residual (ln(g))",
  set_intercept = NULL,
  write_plot = TRUE
)
}
\arguments{
\item{x}{Input data.frame}

\item{region}{Region. AI, BS, or GOA. Character (1L).}

\item{var_x_name}{Name of the x (time variable). Character (1L).}

\item{var_y_name}{Name of the y variable in the data.frame. Character (1L).}

\item{var_y_se_name}{Name of the standard error for the y variable. Character (1L).}

\item{y_title}{Y-axis title. Character (1L).}

\item{set_intercept}{Intercept to use for plots. Numeric (1L). 0 for residual, 1 for VAST relative condition, NULL for VAST a}

\item{write_plot}{Should plots be written to the /plot/ directory?}

\item{fill_colors}{Fill color to use for bars. Character (1L).}
}
\value{
A list of bar plots as ggplot objects
}
\description{
Make condition time series bar plots for single species.
}
\examples{
# EBS single species stratum bar plots

names(EBS_INDICATOR$STRATUM)

plot_species_stratum_bar(
x = EBS_INDICATOR$STRATUM,
region = "BS",
var_x_name = "year",
var_y_name = "stratum_resid_mean",
var_y_se_name = "stratum_resid_se",
y_title = "Length-weight residual (ln(g))",
write_plot = FALSE)
}
