% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{plot_stratum_stacked_bar}
\alias{plot_stratum_stacked_bar}
\title{Generate condition time series plot}
\usage{
plot_stratum_stacked_bar(
  x,
  region,
  fill_palette = "BrBG",
  var_y_name = "mean_wt_resid",
  var_x_name = "year",
  var_group_name = "stratum",
  fill_title = "Stratum",
  y_title = "Length-weight residual (ln(g))"
)
}
\arguments{
\item{x}{Input data.frame}

\item{region}{Region. AI, BS, or GOA. Character (1L).}

\item{var_y_name}{Name of the y variable in the data.frame. Character (1L).}

\item{var_x_name}{Name of the x (time variable). Character (1L).}

\item{var_group_name}{Name of the variable to use for grouping (i.e., stratum). Character (1L).}

\item{y_title}{Y-axis title. Character (1L).}

\item{var_y_se_name}{Name of the standard error for the y variable. Character (1L).}
}
\value{
A ggplot object of the time series
}
\description{
Generate condition time series plot
}
