% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{plot_two_timeseries}
\alias{plot_two_timeseries}
\title{Two condition time series and correlation}
\usage{
plot_two_timeseries(
  x_1,
  x_2,
  region,
  series_name_1 = "Historical",
  var_y_name_1 = "mean_wt_resid",
  var_x_name_1 = "year",
  series_name_2 = "VAST",
  var_y_name_2 = "mean_wt_resid",
  var_x_name_2 = "year",
  y_title = "Condition (Z-score)",
  fill_title = "Method",
  scale_y = TRUE,
  year_break = NULL,
  fill_colors = c("#001743", "#0085CA"),
  format_for = "rmd"
)
}
\arguments{
\item{x_1}{A data.frame for the first time series.}

\item{x_2}{A data.frame for the second time series.}

\item{region}{Region. AI, BS, or GOA. Character (1L).}

\item{var_y_name_1}{Name of the y variable in the data.frame. Character (1L).}

\item{var_x_name_1}{Name of the x (time variable). Character (1L).}

\item{var_y_name_2}{Name of the y variable in the data.frame. Character (1L).}

\item{var_x_name_2}{Name of the x (time variable). Character (1L).}

\item{y_title}{Y-axis title. Character (1L).}

\item{fill_title}{Name of the fill variable to use for plotting. Character (1L).}

\item{scale_y}{Logical. Should y variables be Z-score transformed?}

\item{fill_colors}{Fill colors to use for points. Character (2L).}

\item{format_for}{"rmd" or "png"}
}
\value{
A ggplot object of the time series
}
\description{
Two condition time series and correlation
}
