% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xy_corr.R
\name{plot_xy_corr}
\alias{plot_xy_corr}
\title{Plot of condition versus condition}
\usage{
plot_xy_corr(
  x_1,
  x_2,
  region,
  series_name_1 = "Historical",
  var_y_name_1 = "mean_wt_resid",
  var_y_se_name_1 = "se_wt_resid",
  var_x_name_1 = "year",
  series_name_2 = "VAST",
  var_y_name_2 = "mean_wt_resid",
  var_y_se_name_2 = "se_wt_resid",
  var_x_name_2 = "year",
  format_for = "rmd"
)
}
\arguments{
\item{x_1}{A data.frame for the first time series.}

\item{x_2}{A data.frame for the second time series.}

\item{region}{Region. AI, BS, or GOA. Character (1L).}

\item{var_y_name_1}{Name of the y variable in the data.frame. Character (1L).}

\item{var_y_se_name_1}{Name of the standard error for the y variable. Character (1L).}

\item{var_x_name_1}{Name of the x (time variable). Character (1L).}

\item{var_y_name_2}{Name of the y variable in the data.frame. Character (1L).}

\item{var_y_se_name_2}{Name of the standard error for the y variable. Character (1L).}

\item{var_x_name_2}{Name of the x (time variable). Character (1L).}

\item{format_for}{"rmd" or "png"}

\item{fill_colors}{Fill colors to use for points. Character (2L).}

\item{scale_y}{Logical. Should y variables be Z-score transformed?}

\item{y_title}{Y-axis title. Character (1L).}
}
\value{
A ggplot object of the time series
}
\description{
Plot of condition versus condition
}
