% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify_goa_hauls.R
\name{poststratify_goa_hauls}
\alias{poststratify_goa_hauls}
\title{Post-stratify GOA hauls based on 2025 design}
\usage{
poststratify_goa_hauls(gapdata, method = "use_2025_strata")
}
\arguments{
\item{gapdata}{gapdata output from `gapindex::get_data()`}

\item{method}{"2025_strata" (default)}
}
\description{
Assign 1990-2023 GOA hauls to 2025 strata that align with NMFS Statistical Area boundaries.
}
