% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_condition_doy.R
\name{run_condition_doy}
\alias{run_condition_doy}
\title{Condition indicator calculated based on day of year}
\usage{
run_condition_doy(
  species_code,
  region = "EBS",
  method = "lm",
  form = log(weight_g) ~ log(length_cm):factor(sex) + yday:factor(year)
)
}
\arguments{
\item{species_code}{Species code as a 1L numeric vector}

\item{region}{Region as a 1L character vector ("EBS", "NBS", "AI", "GOA")}

\item{method}{Regression function method, either lm or stan_lm}

\item{form}{Formula for the regression.}
}
\description{
Condition indicator calculated based on day of year
}
