% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_species}
\alias{select_species}
\title{Select species}
\usage{
select_species(species_code, region)
}
\arguments{
\item{species_code}{RACE species code as a 1L numeric vector}

\item{region}{Region as a character vector. One of "GOA", "AI", "NBS", "EBS")}
}
\description{
Select length, weight, and CPUE (kg/km2) from csv files based on region and species code.
}
