% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{set_stratum_order}
\alias{set_stratum_order}
\title{Set stratum plotting order}
\usage{
set_stratum_order(STRATUM, REGION)
}
\arguments{
\item{REGION}{Character vector of length one indicating whether the region is AI, GOA, or BS}

\item{COMMON_NAME}{Vector of stratum names}
}
\description{
Set stratum plotting order
}
