% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sql_to_rqry}
\alias{sql_to_rqry}
\title{Retrieve SQL query string from .sql files}
\usage{
sql_to_rqry(sql_path)
}
\arguments{
\item{sql_path}{File path to .sql file as a character vector.}
}
\description{
This function extracts a query strings from simple .sql file that only contain a comment block describing an SQL query and the query itself.
}
