% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groundfish_Condition.R
\name{weighted_resids}
\alias{weighted_resids}
\title{A function to weight length-weight residuals by catch}
\usage{
weighted_resids(year, residuals, catch = 1)
}
\arguments{
\item{year}{Year of sample must be the same length as the residuals}

\item{catch}{Catch for weighting residual (default = 1) must be the same length as residuals}

\item{residual}{Residual that will be weighted by catch}
}
\description{
This function weights length-weight residuals by a catch column. This
catch can be CPUE from the tow where the fish was caught (most common) or
stratum CPUE or biomass.
}
\examples{
weighted_resids()
}
\keyword{condition}
\keyword{groundfish}
\keyword{length,}
\keyword{weight,}
