% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ESR_SETTINGS}
\alias{ESR_SETTINGS}
\title{ESR settings}
\format{
A list containing two data.frames (list of species for each region, VAST settings)
}
\usage{
ESR_SETTINGS
}
\description{
ESR species and VAST settings
}
\keyword{datasets}
