% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_condition_data.R
\name{get_condition_data}
\alias{get_condition_data}
\title{Get data for index calculations}
\usage{
get_condition_data(
  channel = NULL,
  species_by_region = NULL,
  subareas_by_region = NULL,
  start_year_by_region = NULL
)
}
\arguments{
\item{channel}{RODBC channel}

\item{species_by_region}{A named list containing numeric vectors with SPECIES_CODES to use for the indicator.}

\item{subareas_by_region}{A named list containing numeric vectors with subarea AREA_ID numbers to use for indicator calculations.}

\item{start_year_by_region}{A named list of numeric vectors indicating the start year for the indicator.}
}
\description{
Retrieve length-weight, and catch data, derive subarea biomass, and write outputs to .csv files in /data/.
}
