% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBS_INDICATOR}
\alias{NBS_INDICATOR}
\title{Northern Bering Sea (2010-2022)}
\format{
A list containing a data frames (indicator for the full region) and a character vector.
\describe{
     \itemize{
       \item FULL_REGION (data frame): Residuals for the full region
         \itemize{
         \item{YEAR}{: Year}
         \item{COMMON_NAME}{: Species common name}
         \item{mean_wt_resid}{: Mean residual for the full region}
         \item{se_wt_resid}{: Standard error of the indicator for the full region}
         \item{vast_condition}{: Estimate of the allometric intercept,a, of the length-weight equation (W = aL^b), estimated using VAST.}
         \item{vast_condition_se}{Standard error of a, estimated using VAST.}
         \item{species_code}{: RACE/GAP species code}
         \item{scaled_vast_condition}{: Anomaly (Z-score) of vast_condition}
         \item{vast_relative_condition}{: Relative condition based on the allometric intercept estimated using VAST, i.e., allometric intercept divided by the time series mean allometric intercept.}
         \item{vast_relative_condition_se}{: Standard error of vast_relative_condition.}
         }
       \item LAST_UPDATE: Date when indicator and data were last updated.
     }
}
}
\source{
\url{https://www.fisheries.noaa.gov/contact/groundfish-assessment-program}
}
\usage{
NBS_INDICATOR
}
\description{
Morphometric condition indicators based on the allometric intercept estimated using VAST and  residuals from a length-weight regression for Pacific cod, walleye pollock (> 250 mm), walleye pollock (100-250 mm), Alaska plaice, and yellowfin sole in the northern Bering Sea.
}
\keyword{datasets}
