% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{plot_anomaly_timeseries}
\alias{plot_anomaly_timeseries}
\title{Generate condition time series plot}
\usage{
plot_anomaly_timeseries(
  x,
  region,
  fill_color = "#0085CA",
  var_y_name = "mean_wt_resid",
  var_y_se_name = "se_wt_resid",
  var_x_name = "year",
  y_title = "Length-weight residual (ln(g))",
  format_for = "rmd"
)
}
\arguments{
\item{x}{Input data.frame}

\item{region}{Region. AI, BS, or GOA. Character (1L).}

\item{fill_color}{Fill color to use for points. Character (1L).}

\item{var_y_name}{Name of the y variable in the data.frame. Character (1L).}

\item{var_y_se_name}{Name of the standard error for the y variable. Character (1L).}

\item{var_x_name}{Name of the x (time variable). Character (1L).}

\item{y_title}{Y-axis title. Character (1L).}

\item{format_for}{"rmd" or "png"}
}
\value{
A ggplot object of the time series
}
\description{
Generate condition time series plot
}
