% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plots.R
\name{plot_species_stratum_bar}
\alias{plot_species_stratum_bar}
\title{Make single species stratum bar plots}
\usage{
plot_species_stratum_bar(
  x,
  region,
  var_x_name,
  var_y_name,
  var_y_se_name,
  y_title = "Length-weight residual (ln(g))",
  var_group_name = "stratum",
  fill_title = "Stratum",
  fill_palette = "BrBG",
  write_plot = TRUE
)
}
\arguments{
\item{x}{Input data.frame}

\item{region}{Region. AI, BS, or GOA. Character (1L).}

\item{var_x_name}{Name of the x (time variable). Character (1L).}

\item{var_y_name}{Name of the y variable in the data.frame. Character (1L).}

\item{var_y_se_name}{Name of the standard error for the y variable. Character (1L).}

\item{y_title}{Y-axis title. Character (1L).}

\item{var_group_name}{Name of the variable to use for grouping (i.e., stratum). Character (1L).}

\item{fill_title}{Name of the fill variable to use for plotting. Character (1L).}

\item{write_plot}{Should plots be written to the /plot/ directory?}

\item{fill_pallete}{Character vector denoting which color pallette to use. Must be a valid name for RColorBrewer::brewer.pal(name = {fill_palette})}
}
\value{
A list of bar plots as ggplot objects
}
\description{
Make condition time series bar plots with separate panels for each stratum.
}
