% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{set_plot_order}
\alias{set_plot_order}
\title{Set species plotting order}
\usage{
set_plot_order(common_name, region)
}
\arguments{
\item{common_name}{Vector of common names}

\item{region}{Character vector of length one indicating whether the region is AI, GOA, or BS}
}
\description{
Set species plotting order
}
