% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_lw_residuals.R
\name{weight_lw_residuals}
\alias{weight_lw_residuals}
\title{A function to weight length-weight residuals by catch}
\usage{
weight_lw_residuals(
  residuals,
  year,
  stratum = NA,
  stratum_biomass = NA,
  catch = 1
)
}
\arguments{
\item{residuals}{Vector of residuals residuals that will be weighted by stratum biomass or catch.}

\item{year}{Vector of years Year of sample must be the same length as the residuals.}

\item{stratum}{Vector of strata for weighting.}

\item{stratum_biomass}{Vector of stratum biomass.}

\item{catch}{Vector of catch for weighting residuals by haul (default = 1).}
}
\description{
This function weights length-weight residuals by stratum biomass OR catch.
}
\keyword{condition.}
\keyword{groundfish}
\keyword{length,}
\keyword{weight,}
